refactor EditableInput
